Allow multiple PeerTube instances by not hard-coding host.
Probably not the 100% most elegant solution, but it's nice and compact.
You can't selectively copy the embed URL from PeerTube, it'll always
auto-select the entire <iframe> block that you'd need to edit.
Well with a little URL re-writing I can make this just figure it out
from the default video view / share URL, which isn't that annoying to copy.